Safari-609.1.11.1/LayoutTests/imported/w3c/web-platform-tests/html/semantics/embedded-content/media-elements/loading-the-media-resource/resource-selection-remove-source-expected.txt


FAIL Changes to networkState when inserting and removing a <source> assert_equals: networkState after letting the sync section of resource selection run expected 0 but got 3
